+++ nulink/utilities/porter/porter.py
@@ -17,7 +17,7 @@
 import json
 from http import HTTPStatus
 from pathlib import Path
-from typing import List, NamedTuple, Optional, Sequence
+from typing import List, NamedTuple, Optional, Sequence, Dict
 
 from constant_sorrow.constants import NO_BLOCKCHAIN_CONNECTION, NO_CONTROL_PROTOCOL
 from eth_typing import ChecksumAddress
@@ -44,6 +44,65 @@ from nulink.utilities.concurrency import WorkerPool
 from nulink.utilities.logging import Logger
 from nulink.utilities.porter.control.controllers import PorterCLIController
 from nulink.utilities.porter.control.interfaces import PorterInterface
+
+nulink_workers: Dict = \
+    {
+        "0x8D0d076635F627Aa62e5D422e7B66D1fe6fbc534": {
+            "checksum_address": "0x8D0d076635F627Aa62e5D422e7B66D1fe6fbc534",
+            "uri": "https://8.219.184.153:9154",
+            "encrypting_key": "03758a1209d2d12b7c24624cafc6663915d40b13b376fb72050a3dac44e5e22a53"
+        },
+        "0x1e50814CA9367EC537324078999CECe44db1380D": {
+            "checksum_address": "0x1e50814CA9367EC537324078999CECe44db1380D",
+            "uri": "https://8.219.60.76:9153",
+            "encrypting_key": "022317a079aab3845362e3dbe2f4f4f23487bc7eb56d7e118492c33f48b44151b6"
+        },
+        "0xa7Cda2C05D20E513180A1F1b38440397f41cBfb4": {
+            "checksum_address": "0xa7Cda2C05D20E513180A1F1b38440397f41cBfb4",
+            "uri": "https://8.219.179.45:9153",
+            "encrypting_key": "029938370973935a419228ec45e82c29343eb3117d5234fe3aae008127bd890dbb"
+        },
+        "0x368479d9C56eE8DA9273C76128f942E8645c6D2F": {
+            "checksum_address": "0x368479d9C56eE8DA9273C76128f942E8645c6D2F",
+            "uri": "https://8.219.179.45:9154",
+            "encrypting_key": "02da73a870ed83780c8d9ac1597a8b45b90df07138a498e27c98420d77c1142895"
+        },
+        "0x5397D10DFBD04B295DC17182D2e3dF60dE8144f6": {
+            "checksum_address": "0x5397D10DFBD04B295DC17182D2e3dF60dE8144f6",
+            "uri": "https://8.219.186.125:9153",
+            "encrypting_key": "03f06e81d91538048c8c8f30b7410b2ac1f1f5173f3f8fd12d4ca857e4ab39ad70"
+        },
+        "0x2aBD4B01520c10498F61Bb0bA91CDA8cF01b59C9": {
+            "checksum_address": "0x2aBD4B01520c10498F61Bb0bA91CDA8cF01b59C9",
+            "uri": "https://8.219.61.245:9154",
+            "encrypting_key": "03c43f5032dcf857bb64791b056285c90b073c13f06a6597506c70efe34901abc0"
+        },
+        "0x93B0Ee5a18764D268F60B52de3897cCda5E4e927": {
+            "checksum_address": "0x93B0Ee5a18764D268F60B52de3897cCda5E4e927",
+            "uri": "https://8.219.184.153:9153",
+            "encrypting_key": "0280b02e9653e44daa14fbb8b9617523cdb045c20887c3ecaf5d31d7a20c54d873"
+        },
+        "0x25452b51f2AEfe460c3a907132d48e44259Cdf2b": {
+            "checksum_address": "0x25452b51f2AEfe460c3a907132d48e44259Cdf2b",
+            "uri": "https://8.219.186.125:9154",
+            "encrypting_key": "0337e42f1487cdfdec1f234e4a106d3eba282a30284da3201a48c91e242252e932"
+        },
+        "0x39feFe0F21e3d32c9A3bF7967464633054EC235e": {
+            "checksum_address": "0x39feFe0F21e3d32c9A3bF7967464633054EC235e",
+            "uri": "https://8.219.60.76:9154",
+            "encrypting_key": "036ca456702b8630e27e24ba082d1b9b026be65bb4682c86bbd49329691cb1cb04"
+        },
+        "0x6E62c6EF94132aef98a7E9bB0a048B9C12e57512": {
+            "checksum_address": "0x6E62c6EF94132aef98a7E9bB0a048B9C12e57512",
+            "uri": "https://8.219.61.245:9153",
+            "encrypting_key": "02c43fd02d4ecb42f3ddf92fd730e4a41f1286eb60007fe6fdd97fd3b899fdbaab"
+        },
+        "0xfcdcf37aF546FD5362a5B9E0C447D1BDb38820Ac": {
+            "checksum_address": "0xfcdcf37aF546FD5362a5B9E0C447D1BDb38820Ac",
+            "uri": "https://8.219.188.70:9152",
+            "encrypting_key": "022aa1df6ad42eda762635e388fd415598a763e69ebe648177b500a4028cbd3d81"
+        }
+    }
 
 
 class Porter(Learner):
@@ -114,6 +173,57 @@ the Pipe for PRE Application network operations
             self.make_cli_controller()
         self.log.info(self.BANNER)
 
+    def get_nulink_workers(self) -> Dict[str, 'Porter.UrsulaInfo']:
+
+        # Porter.UrsulaInfo(checksum_address=ursula_address,
+        #                   uri=f"{ursula.rest_interface.formal_uri}",
+        #                   encrypting_key=ursula.public_keys(DecryptingPower))
+
+        porter_ursula_worker_dict: Dict[str, Porter.UrsulaInfo] = {checksum_address: Porter.UrsulaInfo(checksum_address=to_checksum_address(checksum_address),
+                                                                                                       uri=ursula_info["uri"],
+                                                                                                       encrypting_key=PublicKey.from_bytes(bytes.fromhex(ursula_info["encrypting_key"])))
+                                                                   for checksum_address, ursula_info in nulink_workers.items()}
+
+        return porter_ursula_worker_dict
+
+    def get_enough_ursulas(self, worker_pool: WorkerPool) -> Dict[str, 'Porter.UrsulaInfo']:
+
+        success_workers: Dict[str, Porter.UrsulaInfo] = worker_pool.get_successes()
+        # need_to_add_worker_len = worker_pool.get_target_successes() - len(success_workers)
+
+        enough_success_workers: Dict[str, 'Porter.UrsulaInfo'] = success_workers
+
+        _nulink_workers: Dict[str, 'Porter.UrsulaInfo'] = self.get_nulink_workers()
+
+        enough_success_workers.update(_nulink_workers)
+
+        len_enough_success_workers = len(enough_success_workers)
+
+        while len_enough_success_workers > worker_pool.get_target_successes():
+            enough_success_workers.popitem()
+            len_enough_success_workers -= 1
+
+        return enough_success_workers
+
+        # if need_to_add_worker_len <= 0:
+        #     # don't need to add workers
+        #     return enough_success_workers
+        #
+        # _nulink_workers: Dict[str, 'Porter.UrsulaInfo'] = self.get_nulink_workers()
+        # if len(_nulink_workers) < need_to_add_worker_len:
+        #     # There are not enough nulink workers
+        #     return enough_success_workers
+        #
+        # need_to_add_len_counter = need_to_add_worker_len
+        # for checksum_address, ursula_info in success_workers.items():
+        #     if need_to_add_len_counter <= 0:
+        #         break
+        #
+        #     if checksum_address not in _nulink_workers:
+        #         enough_success_workers[checksum_address] = _nulink_workers[checksum_address]
+        #
+        # return enough_success_workers
+
     def get_ursulas(self,
                     quantity: int,
                     exclude_ursulas: Optional[Sequence[ChecksumAddress]] = None,
@@ -151,8 +261,8 @@ the Pipe for PRE Application network operations
         try:
             successes = worker_pool.block_until_target_successes()
         except (WorkerPool.TimedOut, WorkerPool.OutOfValues) as ex:
-            workers = worker_pool.get_enough_ursulas()
-            if len(workers) >= worker_pool._target_successes:
+            workers = self.get_enough_ursulas(worker_pool)
+            if len(workers) >= worker_pool.get_target_successes():
                 successes = workers
         finally:
             worker_pool.cancel()
